use check_estimator
