+++ src/aiod/contracts/sklearn/__init__.py
@@ -1,4 +1,4 @@
-"""Sklearn models."""
+"""Base class for scikit-learn API contracts."""
 
 from aiod.contracts.sklearn.classification import SklearnClassifierContract
